Remove "ghosts" caused by broadcasts.
The response to a 'get_gamestate' command was being broadcasted to all
clients, not just the requestor.

We now only send 'set_gamestate' to the requestor.

When a new player arrives, we send 'create_entity' before that player
connects, so that they do not have a ghost of themselves on screen.